Add initial implementation of CapsuleCRM MCP extension as Claude DTX with server, API tools, and documentation
