+++ CHANGELOG.md
@@ -1,0 +1,54 @@
+# Changelog
+
+All notable changes to the CapsuleCRM MCP Extension will be documented in this file.
+
+## [1.0.0] - 2025-07-10
+
+### 🎉 Initial Release
+
+#### ✨ Features
+- **Complete CRM Integration**: Full support for CapsuleCRM parties, opportunities, tasks, and milestones
+- **Natural Language Interface**: Talk to your CRM using plain English commands
+- **Advanced Search & Filtering**: Support for 8+ operators and complex queries
+- **Smart Value Calculations**: Automatic probability-weighted opportunity values
+- **Real-time Data Access**: Direct API integration with no data storage
+
+#### 🎯 Core Capabilities
+- **Party Management**: List, create, update, search people and organizations
+- **Opportunity Management**: Full sales pipeline management with advanced filtering
+- **Task Management**: Create, track, and manage tasks across teams
+- **Pipeline Insights**: View milestones and track deal progress
+
+#### 🔧 Technical Features
+- **Production-Ready**: Comprehensive error handling and logging
+- **Bundled Dependencies**: Self-contained extension with all required packages
+- **MCP Protocol**: Full stdio transport compliance for Claude Desktop
+- **Security**: Secure API token handling and validation
+
+#### 📚 Documentation
+- Comprehensive README with usage examples
+- Detailed installation guide
+- Advanced usage patterns and filtering guide
+- Troubleshooting documentation
+
+#### 🚀 Sample Prompts
+- Customer Insights: "Show me all VIP customers from last month"
+- Sales Pipeline Review: "What's the total value of our open opportunities?"
+- Task Management: "What tasks are overdue?"
+- Advanced Filtering: Complex multi-condition searches
+
+### 🛡️ Security & Privacy
+- No data storage - real-time API bridge only
+- Secure environment variable handling
+- User-centric design with no internal ID exposure
+- Official CapsuleCRM API integration
+
+### 🔗 Integration
+- Claude Desktop Extension (DXT) format
+- One-click installation
+- User configuration for API key
+- Cross-platform compatibility
+
+---
+
+*This release represents a complete, production-ready CRM automation solution for Claude AI users.*
