--- main.py
@@ -1,2 +1,0 @@
-# Die FastMCP-Server-Logik befindet sich jetzt in mcp_server.py
-# Bitte dort starten: python mcp_server.py
